Owner configuration cleanup
It should be a lot easier to change how the owner model is associated with the approval model now.

`ActsAsApprovable::Ownership.configure(Approval, User)` to configure. Accepts a block as well.
`.available_owners` should be overridden with an AREL query selecting valid owners.
`.option_for_owner(owner)` is used to turn the owner object into a value array for `options_for_select`.